AROSTCP/bsdsocket: Codacy quality sweep and 64-bit/portability fixes
Quality-only cleanup across netinet, netinet6, net, kern and api:
- cast printf/log %ld/%lx/%lu args to C (long)/(u_long) to match the
  kernel vcsprintf reading long/u_long on 64-bit (LONG/ULONG kept for
  declarations and return values)
- remove dead stores / unused locals, narrow variable scope
- if.c: guard __builtin_offsetof (#if __GNUC__), else standard offsetof
  via <stddef.h> instead of the &((t*)0)->m null-deref trick

Bug fix:
- if_sana.c sana_down(): initialise success (was read uninitialised when
  CreateIOSana2Req() fails)

+++ workbench/network/stacks/AROSTCP/bsdsocket/api/amiga_generic.c
@@ -297,7 +297,7 @@ LONG __IoctlSocket(LONG fdes, ULONG cmd, caddr_t data, struct SocketBase *libPtr
 
 Return:
     ReleaseSyscallSemaphore(libPtr);
-    DOPTERR(if(error) log(LOG_ERR, "IoctlSocket(): error %ld on command 0x%08lx", error, cmd);)
+    DOPTERR(if(error) log(LOG_ERR, "IoctlSocket(): error %ld on command 0x%08lx", (long)error, cmd);)
         API_STD_RETURN(error, 0);
 }
